diacritical/requirements.txt
flask==3.0.2
python-docx==1.1.0
gunicorn==21.2.0
flask-bootstrap==3.3.7.1
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-dotenv==1.0.0 